+++ R/DataMan.R
@@ -1,0 +1,8 @@
+#' DataMan a data clearning, analysis and statistical modeling assessing tool
+#' 
+#' @description a data clearning, analysis and statistical modeling assessing tool.
+#' @docType package
+#' @name DataMan
+#' @useDynLib DataMan
+#' @importFrom Rcpp evalCpp
+NULL